PR Update
Here I have updated the pull request according to the requirments from the repository owner
